+++ tests/monitor/test-jobs.R
@@ -2,7 +2,7 @@ library(crew.aws.batch)
 library(testthat)
 
 test_that("empty job list", {
-  x <- crew_aws_batch_monitor(
+  x <- crew_monitor_aws_batch(
     job_definition = "never-existed",
     job_queue = "crew-aws-batch-job-queue",
     region = "us-east-2"
@@ -13,7 +13,7 @@ test_that("empty job list", {
 })
 
 test_that("job list", {
-  x <- crew_aws_batch_monitor(
+  x <- crew_monitor_aws_batch(
     job_definition = "crew-aws-batch-test",
     job_queue = "crew-aws-batch-job-queue",
     region = "us-east-2"
@@ -86,7 +86,7 @@ test_that("job list", {
 })
 
 test_that("job logs", {
-  x <- crew_aws_batch_monitor(
+  x <- crew_monitor_aws_batch(
     job_definition = "crew-aws-batch-test",
     job_queue = "crew-aws-batch-job-queue",
     region = "us-east-2"