+++ NEWS.md
@@ -1,6 +1,7 @@
 # crew.cluster 0.0.1.9000
 
 * Use `paws.common::paginate()` to get the full log of a job (#5). Requires `paws.common` >= 0.7.0 due to https://github.com/paws-r/paws/issues/721. 
+* Rename `crew_aws_batch_monitor()` to `crew_monitor_aws_batch()` for syntactic consistency.
 
 # crew.cluster 0.0.1